requirements.txt
# Core dependencies for Japanese video subtitle generation and translation
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0

# Translation
transformers>=4.35.0
sentencepiece>=0.1.99
protobuf>=3.20.0
bitsandbytes>=0.41.0
accelerate>=0.20.0

# Video/Audio processing
ffmpeg-python>=0.2.0

# Utilities
tqdm>=4.65.0
numpy>=1.24.0
requests>=2.31.0
